Add integration test in separate file
